:recycle: move concurrency into toolkit
