+++ caikit/core/model_management/local_model_trainer.py
@@ -35,8 +35,8 @@ from ..exceptions import error_handler
 from ..modules import ModuleBase
 from ..toolkit.logging import configure as configure_logging
 from .model_trainer_base import ModelTrainerBase, TrainingInfo
-from caikit.core.concurrency.destroyable_process import DestroyableProcess
-from caikit.core.concurrency.destroyable_thread import DestroyableThread
+from caikit.core.toolkit.concurrency.destroyable_process import DestroyableProcess
+from caikit.core.toolkit.concurrency.destroyable_thread import DestroyableThread
 from caikit.core.exceptions.caikit_core_exception import (
     CaikitCoreException,
     CaikitCoreStatusCode,
