+++ caikit/core/toolkit/concurrency/destroyable_thread.py
